fix: electron not found in prod (#67)
