+++ main/npm/registry/getCurrentRegistry.ts
@@ -1,0 +1,7 @@
+import { NPM_REGISTRY, REGISTRY_FIELD } from '../../constants';
+import { getNpmInfo } from '../npmInfo';
+
+export default async function getCurrentRegistry() {
+  const npmrc = await getNpmInfo();
+  return npmrc[REGISTRY_FIELD] || NPM_REGISTRY;
+}
