+++ main/npm/registry/index.ts
@@ -1,0 +1,3 @@
+export { default as getAllRegistries } from './getAllRegistries';
+export { default as getCurrentRegistry } from './getCurrentRegistry';
+export { default as setCurrentRegistry } from './setCurrentRegistry';
