+++ main/utils/getNpmRegistry.ts
@@ -1,5 +1,5 @@
 import { ALI_NPM_REGISTRY } from '@appworks/constant';
-import { getCurrentRegistry } from '../npm/registry';
+import getCurrentRegistry from '../npm/registry/getCurrentRegistry';
 import checkIsAliInternal from './checkIsAliInternal';
 
 async function getNpmRegistry() {
